Tarea TCC/ClasesEquivalencia/src/test/java/edu/unisabana/dyas/tdd/registry/RegistryTest.java
package edu.unisabana.dyas.tdd.registry;

import org.junit.Assert;
import org.junit.Test;

public class RegistryTest {
    private Registry registry = new Registry();
    @Test
    public void testValidVoter() {
        Person person = new Person("John Doe", 1, 30, Gender.MALE, true);
        RegisterResult result = registry.registerVoter(person);
        Assert.assertEquals(RegisterResult.VALID, result);
    }
    @Test
    public void testUnderageVoter() {
        Person person = new Person("Jane Doe", 2, 16, Gender.FEMALE, true);
        RegisterResult result = registry.registerVoter(person);
        Assert.assertEquals(RegisterResult.UNDERAGE, result);
    }
    @Test
    public void testInvalidAgeVoter() {
        Person person = new Person("Old Man", 3, 130, Gender.MALE, true);
        RegisterResult result = registry.registerVoter(person);
        Assert.assertEquals(RegisterResult.INVALID_AGE, result);
    }
    @Test
    public void testDeadVoter() {
        Person person = new Person("Deceased", 4, 40, Gender.FEMALE, false);
        RegisterResult result = registry.registerVoter(person);
        Assert.assertEquals(RegisterResult.DEAD, result);
    }
    @Test
    public void testDuplicatedVoter() {
        Person person = new Person("Duplicate", 5, 25, Gender.UNIDENTIFIED, true);
        registry.registerVoter(person); // Register once
        RegisterResult result = registry.registerVoter(person); // Attempt to register again
        Assert.assertEquals(RegisterResult.DUPLICATED, result);
    }
    // TODO Complete with more test cases
}
